Implemented form

+++ src/main/java/com/todo/list/repository/TodoListRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.todo.list.model.Todo;
 
-public interface TodoListRepository extends CrudRepository<Todo, Double> {
+public interface TodoListRepository extends CrudRepository<Todo, Long> {
 }
